Sticky acks, comment and downtime updates, notification updates, new external commands
git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@273 5f96b256-904b-4d8d-8c98-d829582c6739

+++ xdata/xcddefault.h
@@ -3,7 +3,7 @@
  * XCDDEFAULT.H - Header file for default comment data routines
  *
  * Copyright (c) 2000-2002 Ethan Galstad (nagios@nagios.org)
- * Last Modified:   03-16-2002
+ * Last Modified:   03-21-2002
  *
  * License:
  *
@@ -34,10 +34,10 @@ int xcddefault_create_comment_file(void);
 int xcddefault_validate_comment_data(void);
 int xcddefault_cleanup_comment_data(char *);
 int xcddefault_save_comment_data(void);
-int xcddefault_add_new_host_comment(char *,time_t,char *,char *,int,int,int *);
-int xcddefault_add_new_service_comment(char *,char *,time_t,char *,char *,int,int,int *);
-int xcddefault_delete_host_comment(int);
-int xcddefault_delete_service_comment(int);
+int xcddefault_add_new_host_comment(char *,time_t,char *,char *,int,int,unsigned long *);
+int xcddefault_add_new_service_comment(char *,char *,time_t,char *,char *,int,int,unsigned long *);
+int xcddefault_delete_host_comment(unsigned long);
+int xcddefault_delete_service_comment(unsigned long);
 int xcddefault_delete_all_host_comments(char *);
 int xcddefault_delete_all_service_comments(char *,char *);
 #endif
